\name{cancer}
\alias{cancer}
\docType{data}
\title{Cancer Survival data}
\description{
Patients with advanced cancers of the stomach, bronchus, colon, ovary or breast were treated with ascorbate. The purpose of the study was to determine if the survival times differ with respect to the organ affected by the cancer.
}
\format{This data frame contains the following columns:
 \describe{
 \item{\code{Survival}}{time in days}
 \item{\code{Organ}}{Organ affected by the cancer}
 }}
\usage{data(cancer)}



\references{
Cameron, E. and Pauling, L. (1978) Supplemental ascorbate in the supportive treatment of cancer: re-evaluation of prolongation of survival times in terminal human cancer. Proceedings of the National Academy of Science USA, 75, 4538-4542. 

Also found in: Manly, B.F.J. (1986) Multivariate Statistical Methods: A Primer, New York: Chapman and Hall, 11. Also found in: Hand, D.J., et al. (1994) A Handbook of Small Data Sets, London: Chapman and Hall, 255.
}
\keyword{datasets}