\name{desaturate_color}
\alias{desaturate_color}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Desaturates colors
}
\description{Desaturates colors}
\usage{
desaturate_color(cols, frac = 0.8)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{cols}{Colors}
  \item{frac}{Fraction to desaturate by.}
}
\value{Desaturated version of orginal colors}
