\name{overview}
\alias{overview}
\alias{PairViz-package}
\docType{package}
\title{ Overview of PairViz package }
\description{
 Implements methods described in Hurley and Oldford paper.
 
 There are functions for constructing eulerian paths on complete graphs- see \code{\link{eseq}}, \code{\link{hpaths}}, and \code{\link{weighted_hpaths}}, and eulerians on general graphs-
 see \code{\link{etour}} and \code{\link{eulerian}}.
 
 There are also functions for new types of graphics, \code{\link{mc_plot}}, \code{\link{catpcp}} and
 	\code{\link{guided_pcp}} and a barchart/mosaic variant \code{\link{table_plot}}.
 	
  
}

\author{ C.B. Hurley and R.W. Oldford }

\references{C.B. Hurley and R.W. Oldford,
Pairwise display of high dimensional information via 
Eulerian tours and Hamiltonian decompositions. Journal of Computational and Graphical Statistics.
19(10), pp. 861--886, 2010.

C.B. Hurley and R.W. Oldford, 
Eulerian tour algorithms for data visualization and the PairViz package. 
Computational Statistics, 26(4), pp 613--633, 2011. 
}

\keyword{graphs}
\keyword{optimize}
\keyword{hplot}
